refactor: Realm MapKit Support
